First draft of lambda function that reattaches eni
For zookeeper (other less cloud native cluster tech) we need the nodes of a cluster
to keep the same IP address. When ec2 instances are autohealed by and autoscaling group
this function looks to reattach an existing network interface that was attached to the
previous instance that this one is replacing.